Introduce a custom error for unknown messages
The custom error with information on how to fix it by trying to find the
closest possible message by Levenshtein distance and falling back to
listing all messages.

Along the way, introduce a standard `inspect` method for instances of
user defined classes and adjusted the "inspect" style of a few objects.